Stop Git changing line endings
The files while mostly use LF line endings some also use LF/CR line endings. This change stops Git changing the line endings so Lint is usable (I.E. it passes all files with no errors on a clean checkout).

Note: It is possible to configure Git to not change line endings globally (this is how AppVeyor is configured so the issue does not show on there) but it is better to set in the repository so not to catch out people.